https Server
iOS 7.1 requires https for the plist and ipa links.
See:
http://stackoverflow.com/questions/20276907/enterprise-app-deployment-do
esnt-work-on-ios-7-1/

I changed the server so it acts as a https server.

However there is still a problem in getting the iOS device to accept
your self signed certificates.
But this has nothing to do with the app (see the above link for
possible solutions)